airflow/requirements.txt
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
cramjam==2.9.0
fastparquet==2024.5.0
fsspec==2024.10.0
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-dataproc==5.13.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.0
grpcio-status==1.67.0
idna==3.10
numpy==2.1.2
packaging==24.1
pandas==2.2.3
proto-plus==1.24.0
protobuf==5.28.2
py4j==0.10.9.7
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyspark==3.5.3
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rsa==4.9
six==1.16.0
soupsieve==2.6
tzdata==2024.2
urllib3==2.2.3
apache-airflow-providers-google==12.0.0
